+++ docs/Whats-New.md
@@ -247,6 +247,7 @@ New:
 - Script Action 110 that Modify how ends the new move actions (by FS-21)
 - Script Action 111 that un-register Team success, is just the opposite effect of Action 49 (by FS-21)
 - Script Action 112 to regroup temporarily around the Team Leader (by FS-21)
+- Script Action 113 to Randomly Skip Next Action (by FS-21)
 - ObjectInfo now shows current Target and AI Trigger data (by FS-21)
 - Shield absorption and passthrough customization (by Morton)
 - Limbo Delivery of buildings (by Morton)